Ensure signed tokens exist during rollout
Until all existing pending signatures have been verified it's possible
that signed_token may be nil so generate it dynamically if it's not
been set. This ensures that anyone following the verify link sees the
signed page and any existing sharing of a signed link with a token in
the url gets redirected to the petition page.