Clear up signature signed tokens
Some people sign a lot of petitions in a single browser session so
remove tokens for signatures that have previously been validated and
use it once on the signed page. This changes the behaviour on refreshing
the signed page to redirect to the petition page but that's probably
a better behaviour anyway.